// src/app/core/types/custom-values-central-image-card.ts
export type CustomValuesCentralImageCard = {
    header: {
        background: string;
        fontColor: string;
    },
    imageContainer: {
        background: string;
    },
    cardTopPart: {
        background: string,
        colorBadges: string;
        badgesColorFont: string;
        labelsColorFont: string;
    },
    cardBottomPart: {
        background: string;
        statColorFont: string;
        progressBarColor: string;
        progressBarBackground: string;
        progressBarColorFont: string;
    }
}